Move the `ParserOptions` type to `ml-ast` from `ml-config`
