+++ packages/@markuplint/ml-config/package.json
@@ -16,6 +16,7 @@
 		"clean": "tsc --build --clean"
 	},
 	"devDependencies": {
+		"@markuplint/ml-ast": "3.0.0-alpha.0",
 		"@types/mustache": "^4.1.2"
 	},
 	"dependencies": {
