+++ packages/@markuplint/ml-config/tsconfig.json
@@ -11,6 +11,9 @@
 	},
 	"references": [
 		{
+			"path": "../ml-ast"
+		},
+		{
 			"path": "../selector"
 		}
 	],
