+++ packages/@markuplint/ml-core/src/types.ts
@@ -1,8 +1,7 @@
 import type { AnyMLRule } from './ml-rule';
 import type Ruleset from './ruleset';
 import type { LocaleSet } from '@markuplint/i18n';
-import type { MLMarkupLanguageParser } from '@markuplint/ml-ast';
-import type { ParserOptions } from '@markuplint/ml-config';
+import type { MLMarkupLanguageParser, ParserOptions } from '@markuplint/ml-ast';
 import type { ExtendedSpec, MLMLSpec } from '@markuplint/ml-spec';
 
 export type MLSchema = Readonly<[MLMLSpec, ...ExtendedSpec[]]>;
